refactor:Refactor: Adjust ImageProcessingService to use interfaces and move implementations to infrastructure
