+++ src/domain/services/IImageProcessingService.ts
@@ -1,0 +1,5 @@
+import { Task } from "../entities/task.entity";
+
+export interface IImageProcessingService {
+  process(task: Task): Promise<{ resolution: string; path: string }[]>;
+}
